Fixing Size_T marshalling
using vcruntime intrinsic to match processor architecture.

+++ DirectX.DXGI/DXGI_ADAPTER_DESC1.h
@@ -15,9 +15,9 @@ namespace DirectX
             unsigned int DeviceId;
             unsigned int SubSysId;
             unsigned int Revision;
-            unsigned long long DedicatedVideoMemory;
-            unsigned long long DedicatedSystemMemory;
-            unsigned long long SharedSystemMemory;
+            size_t DedicatedVideoMemory;
+            size_t DedicatedSystemMemory;
+            size_t SharedSystemMemory;
             LUID AdapterLuid;
             unsigned int Flags;
         };
